Extract quickstarts banner into its own component
